+++ packages/components/src/Cluster/Cluster.module.css.d.ts
@@ -1,0 +1,10 @@
+declare const styles: {
+  readonly "cluster": string;
+  readonly "sm": string;
+  readonly "md": string;
+  readonly "lg": string;
+  readonly "xl": string;
+  readonly "collapsed": string;
+};
+export = styles;
+