+++ packages/components/src/Tiles/Tiles.module.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "tiles": string;
+};
+export = styles;
+