+++ packages/site/src/layout/Layout.tsx
@@ -6,6 +6,8 @@ import "./code-theme.css";
 import { hooksList } from "../hooksList";
 import { NotFoundPage } from "../pages/NotFoundPage";
 import { TritonSideDrawer } from "../components/TritonSideDrawer";
+import { VisualTestComponentPage } from "../pages/VisualTestComponentPage";
+import { VisualTestLayoutPage } from "../pages/VisualTestLayoutPage";
 
 /**
  * Layout for whole application. This will display the NavMenu and the content of the page.
@@ -15,24 +17,16 @@ import { TritonSideDrawer } from "../components/TritonSideDrawer";
 export const Layout = () => {
   const location = useLocation();
   const scrollPane = useRef<HTMLDivElement>(null);
-  const path = new URLSearchParams(location.search).get("path");
-  const history = useHistory();
   useEffect(() => {
     if (scrollPane?.current) {
       scrollPane?.current.scrollTo({ top: 0 });
     }
   }, [location, scrollPane?.current]);
 
-  // Redirects for the links on the hooks packages page
-  if (path && path.includes("hooks")) {
-    const pathRegex = /hooks-(.*)--docs/g.exec(path);
-    const match = hooksList.find(
-      hook => pathRegex?.[1] === hook.title.toLowerCase(),
-    );
+  useHookRedirect();
 
-    if (match) {
-      history.push(match.to);
-    }
+  if (location.pathname.includes("visual-tests")) {
+    return <OutOfLayoutSwitch />;
   }
 
   return (
@@ -52,6 +46,39 @@ export const Layout = () => {
       </div>
       <TritonSideDrawer />
     </LayoutWrapper>
+  );
+};
+
+const useHookRedirect = () => {
+  const path = new URLSearchParams(location.search).get("path");
+  const history = useHistory();
+
+  if (path && path.includes("hooks")) {
+    const pathRegex = /hooks-(.*)--docs/g.exec(path);
+    const match = hooksList.find(
+      hook => pathRegex?.[1] === hook.title.toLowerCase(),
+    );
+
+    if (match) {
+      history.push(match.to);
+    }
+  }
+};
+
+const OutOfLayoutSwitch = () => {
+  return (
+    <Switch>
+      <Route
+        exact={true}
+        path="/visual-tests/components"
+        component={VisualTestComponentPage}
+      />
+      <Route
+        exact={true}
+        path="/visual-tests/layout"
+        component={VisualTestLayoutPage}
+      />
+    </Switch>
   );
 };