+++ packages/site/src/maps/guidesContent.tsx
@@ -5,6 +5,8 @@ import FrontendStyleguideComponent from "@atlantis/docs/guides/frontend-style.st
 import AddingAnIcon from "@atlantis/docs/guides/adding-an-icon.stories.mdx";
 import GettingStartedWithReactComponent from "@atlantis/docs/getting-started-with-react/getting-started-with-react.stories.mdx";
 import PullRequestTitleGeneratorComponent from "@atlantis/docs/guides/pull-request-title-generator.stories.mdx";
+import PageLayoutsComponent from "../guides/page-layouts.stories.mdx";
+import ScaffoldingComponent from "../guides/scaffolding.stories.mdx";
 import { ContentMapItems } from "../types/maps";
 
 export const guidesContentMap: ContentMapItems = {
@@ -42,5 +44,16 @@ export const guidesContentMap: ContentMapItems = {
     intro: "Pull request title generator",
     title: "Pull request title generator",
     content: () => <PullRequestTitleGeneratorComponent />,
+  },
+  "page-layouts": {
+    intro: "Page Layouts",
+    title: "Page Layouts",
+    content: () => <PageLayoutsComponent />,
+  },
+  scaffolding: {
+    intro: "Scaffolding",
+    title: "Scaffolding",
+    content: () => <ScaffoldingComponent />,
+    noMaxWidth: true,
   },
 };